fixed revelles algorithm, added device lambda XYZtoId support

+++ include/octree.cuh
@@ -131,6 +131,6 @@ __device__ unsigned char raycastDrawPixel(Octree* octree, float oX, float oY, fl
 
 __device__ int proc_subtree(Octree* octree, float oX, float oY, float oZ, float dX, float dY, float dZ, float tx0, float ty0, float tz0, float tx1, float ty1, float tz1, unsigned char a, int minNodeSize, int sX, int sY, uchar4* pixels, int morton = 1);
 
-__device__ int traverseChildNodes(Stack::Frame* data, unsigned char a, int minNodeSize, int sX, int sY, int origOX, int origOY, int origOZ, float origDX, float origDY, float origDZ, uchar4* pixels, Stack& stack, Octree* octree);
+__device__ int traverseChildNodes(Stack::Frame* data, unsigned char a, int minNodeSize, int sX, int sY, float origOX, float origOY, float origOZ, float origDX, float origDY, float origDZ, uchar4* pixels, Stack& stack, Octree* octree);
 
-__device__ int traverseNewNode(float tx0, float ty0, float tz0, float&tx1, float ty1, float tz1, unsigned int nodeIdx, int minNodeSize, int sX, int sY, int origOX, int origOY, int origOZ, float origDX, float origDY, float origDZ, uchar4* pixels, Stack& stack, Octree* octree);
+__device__ int traverseNewNode(float tx0, float ty0, float tz0, float&tx1, float ty1, float tz1, unsigned int nodeIdx, int minNodeSize, int sX, int sY, float origOX, float origOY, float origOZ, float origDX, float origDY, float origDZ, uchar4* pixels, Stack& stack, Octree* octree);
